Moved display components to images
