+++ src/app/features/events/EventList/CalendarComponents/CalendarAgendaEvent.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types'
 
 import { withTheme } from '@material-ui/core'
 
-import WeaponDisplay from 'app/shared/display/WeaponDisplay'
+import WeaponDisplay from 'app/shared/images/WeaponDisplay'
 import useIsMobileMode from 'app/shared/hooks/useIsMobileMode'
 
 function CalendarAgendaEvent ({ event, theme }) {
