+++ src/app/shared/images/LoadoutItem/LoadoutItem.css
