+++ src/app/shared/images/LoadoutItem/LoadoutItem.js
