+++ src/app/shared/images/LoadoutItem/LoadoutItemImage.js
