+++ src/app/shared/images/LoadoutItem/index.js
