+++ src/app/shared/images/WeaponDisplay.css
