+++ src/app/shared/images/WeaponDisplay.js
